Change sentence storage method
Data about sentences is now stored inside binary strings. More compact than how it worked before, and more information is stored.
Data is split into buckets so the performances is reasonable. Each bucket contains a sequence of 11 byte blocks containing the sentence information.
First 5 bytes are the key of the sentence within the bucket, Next 4 bytes are card id, Last 2 bytes are index of sentence in card.
Still uses the one pass algorithm from the last implementation, but this method of storage is more flexible and allows for better algorithms.