Fix bug in pgsync
Current version of pgsync has a bug that prevents grandchild nodes from having through_tables, which is needed for indexing tags to work
The fix is pretty simple, so for now we can just copy fixed versions of the relevant files to the container.
Hopefully I will submit a pull request at some point fixing the issue.